feat: completed main app functions
I added code to check if the input file is a video or an image, perform the preprocessing accordingly and exit with a warning message if the input file is not a video or an image.

I created functions for drawing bounding boxes on detected persons and counting total persons in a frame. I made the bounding box such that labels and confidence values are written at the top of the bounding box. I also added code to calculate total persons in the frame and the duration the person stays in the frame by using results from the function that counts the total persons in a frame.

I added code to stream the output video to the web and save the output image to a folder.

Resolves: #100, #200