chore: Crucial updates
I  added the XML file of the Intermediate representation of the model. The BIN file could not be uploaded as well because of the 25MB file transfer size limit imposed by GitHub. The BIN will be uploaded later as soon a means to accomplish this task has been discovered

I also updated README.md with important information that was previously omitted from the document.